// lambdas/lib/get-triggers/src/types/GetTriggersResponse.ts
export const AaveStopLossToCollateralV2ID = 111n as const
export const AaveStopLossToDebtV2ID = 112n as const
export const SparkStopLossToCollateralV2ID = 117n as const
export const SparkStopLossToDebtV2ID = 118n as const
export const AaveBasicBuyV2ID = 119n as const
export const AaveBasicSellV2ID = 120n as const

export type AaveStopLossToCollateral = {
  triggerTypeName: 'AaveStopLossToCollateralV2'
  triggerType: typeof AaveStopLossToCollateralV2ID
  triggerId: string
  decodedParams: {
    positionAddress: string
    triggerType: string
    maxCoverage: string
    debtToken: string
    collateralToken: string
    ltv: string
  }
}

export type AaveStopLossToDebt = {
  triggerTypeName: 'AaveStopLossToDebtV2'
  triggerType: typeof AaveStopLossToDebtV2ID
  triggerId: string
  decodedParams: {
    positionAddress: string
    triggerType: string
    maxCoverage: string
    debtToken: string
    collateralToken: string
    ltv: string
  }
}

export type SparkStopLossToCollateral = {
  triggerTypeName: 'SparkStopLossToCollateralV2'
  triggerType: typeof SparkStopLossToCollateralV2ID
  triggerId: string
  decodedParams: {
    positionAddress: string
    triggerType: string
    maxCoverage: string
    debtToken: string
    collateralToken: string
    ltv: string
  }
}

export type SparkStopLossToDebt = {
  triggerTypeName: 'SparkStopLossToDebtV2'
  triggerType: typeof SparkStopLossToDebtV2ID
  triggerId: string
  decodedParams: {
    positionAddress: string
    triggerType: string
    maxCoverage: string
    debtToken: string
    collateralToken: string
    ltv: string
  }
}

export type AaveBasicBuy = {
  triggerTypeName: 'AaveBasicBuyV2'
  triggerType: typeof AaveBasicBuyV2ID
  triggerId: string
  decodedParams: {
    positionAddress: string
    triggerType: string
    maxCoverage: string
    debtToken: string
    collateralToken: string
    opHash: string
    execLtv: string
    targetLtv: string
    maxBuyPrice: string
    deviation: string
    maxBaseFeeInGwei: string
  }
}

export type AaveBasicSell = {
  triggerTypeName: 'AaveBasicSellV2'
  triggerType: typeof AaveBasicSellV2ID
  triggerId: string
  decodedParams: {
    positionAddress: string
    triggerType: string
    maxCoverage: string
    debtToken: string
    collateralToken: string
    opHash: string
    execLtv: string
    targetLtv: string
    minSellPrice: string
    deviation: string
    maxBaseFeeInGwei: string
  }
}

export type GetTriggersResponse = {
  triggers: {
    aaveStopLossToCollateral?: AaveStopLossToCollateral
    aaveStopLossToDebt?: AaveStopLossToDebt
    sparkStopLossToCollateral?: SparkStopLossToCollateral
    sparkStopLossToDebt?: SparkStopLossToDebt
    aaveBasicBuy?: AaveBasicBuy
    aaveBasicSell?: AaveBasicSell
  }
}